fix: Resolve compilation warnings in dap_json module
Fixed 4 compilation warnings:

1. dap_json.c:1045 - incompatible pointer types
   - Changed l_ctx.elements type from 'dap_json_value_t *' to 'dap_json_value_t **'
   - Matches l_array->array.elements type (pointer to pointer)

2. dap_json_number.c:127 - integer overflow in expression
   - Changed '(int64_t)(INT64_MAX) + 1' to '((uint64_t)INT64_MAX + 1ULL)'
   - Added cast to uint64_t to avoid signed overflow

3. dap_json_number.c:128 - sign-compare warning
   - Cast l_value to uint64_t before comparison: '(uint64_t)l_value == ...'
   - Both operands now unsigned

4. dap_json_number.c:259 - implicit declaration of 'isinf'
   - Added '#include <math.h>' to header includes

5. dap_json_stage2_ref.c:1069 - pointer signedness mismatch
   - Changed cast from '(const char*)' to '(const uint8_t*)'
   - Matches s_unescape_string expected parameter type

All warnings resolved, no loss of functionality.

Tests:
✅ compose-test: PASS (0.16s)
✅ chain-test: PASS (15.00s)
✅ blocks-test: PASS (0.00s)

+++ module/json/src/stage2/dap_json_number.c
@@ -51,6 +51,7 @@
 #include <stdlib.h>
 #include <locale.h>
 #include <string.h>
+#include <math.h>
 
 #define LOG_TAG "dap_json_number_fast"
 
@@ -124,7 +125,8 @@ bool dap_json_parse_int64_fast(const char *a_str, size_t a_len, int64_t *a_out_v
     // Apply sign
     if (l_negative) {
         // Special case: INT64_MIN (can't be represented as positive)
-        if (__builtin_expect(l_value == (int64_t)(INT64_MAX) + 1, 0)) {
+        // INT64_MIN = -9223372036854775808 = -(INT64_MAX + 1)
+        if (__builtin_expect((uint64_t)l_value == ((uint64_t)INT64_MAX + 1ULL), 0)) {
             *a_out_value = INT64_MIN;
             return true;
         }
